moved non-platform code to a new file
